Create README.md
I've recently completed an engaging and dynamic web project using React.js that revolves around a collection of 100 unique and entertaining memes sourced through an API. This project showcases my proficiency in front-end development and my ability to integrate external data sources seamlessly into a web application.

The heart of the project lies in its dynamic nature, where the memes are fetched in real-time from the API, ensuring that users are always presented with a fresh and diverse selection of humor. The user interface is designed with a modern and intuitive layout, allowing visitors to effortlessly scroll through the extensive meme collection and enjoy a constant stream of laughter.